requirements.txt
elevenlabs[pyaudio]==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
sounddevice==0.4.6
scipy==1.11.1
openai==1.12.0
httpx==0.24.1
flask==3.0.0
flask-socketio==5.3.6
gunicorn==21.2.0
eventlet==0.33.3 